lyrr/src/requirements.txt
requests
aiohttps
lyricgenius
bs4
numpy
datasets
transformers
